// telegram-bot/gmx-sdk/gmx/domain/synthetics/trade/types.ts
export enum TradeActionType {
  OrderCreated = "OrderCreated",
  OrderExecuted = "OrderExecuted",
  OrderCancelled = "OrderCancelled",
  OrderUpdated = "OrderUpdated",
  OrderFrozen = "OrderFrozen",
}

export type PositionTradeAction = {
  marketAddress: string;
  indexToken: TokenData;
  initialCollateralToken: TokenData;
  initialCollateralTokenAddress: string;
  initialCollateralDeltaAmount: BigNumber;
  sizeDeltaUsd: BigNumber;
  isLong: boolean;
  orderType: OrderType;
  triggerPrice?: BigNumber;
  executionPrice?: BigNumber;
  eventName: TradeActionType;
  acceptablePrice: BigNumber;
  indexTokenPriceMax?: BigNumber;
  indexTokenPriceMin?: BigNumber;
};